Auto mode with a toggle button
