+++ reactcode/src/components/toggleButton/toggleButton.css
@@ -1,0 +1,39 @@
+.toggle-button {
+    display: inline-block;
+    cursor: pointer;
+    width: 60px;
+    height: 30px;
+    border-radius: 15px;
+    background-color: #ccc;
+    position: relative;
+    transition: background-color 0.3s ease-in-out;
+  }
+  
+  .toggle-button::before {
+    content: '';
+    position: absolute;
+    top: 2px;
+    left: 2px;
+    width: 26px;
+    height: 26px;
+    border-radius: 50%;
+    background-color: #fff;
+    transition: transform 0.3s ease-in-out;
+  }
+  
+  .toggle-button.on {
+    background-color: #66bb6a;
+  }
+  
+  .toggle-button.on::before {
+    transform: translateX(30px);
+  }
+  
+  .toggle-button.off {
+    background-color: #e0e0e0;
+  }
+  
+  .toggle-button.off::before {
+    transform: translateX(2px);
+  }
+
