--- reactcode/src/pages/autoAnnotate/App.css
@@ -1,49 +1,0 @@
-.slider-container {
-    width: 300px;
-    display: flex;
-    align-items: center;
-  }
-  
-  .custom-slider {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 100%;
-    height: 5px;
-    background: linear-gradient(to right, #4c9aff, #b98aff);
-    outline: none;
-    border-radius: 5px;
-    opacity: 0.7;
-    transition: opacity 0.2s;
-    cursor: pointer;
-  }
-  
-  .custom-slider::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 16px;
-    height: 16px;
-    border-radius: 50%;
-    background: #4c9aff;
-    cursor: pointer;
-    transition: background-color 0.2s;
-  }
-  
-  .custom-slider::-moz-range-thumb {
-    width: 16px;
-    height: 16px;
-    border-radius: 50%;
-    background: #4c9aff;
-    cursor: pointer;
-    transition: background-color 0.2s;
-  }
-  
-  .custom-slider::-webkit-slider-thumb:hover,
-  .custom-slider::-moz-range-thumb:hover {
-    background: #2c7dff;
-  }
-  
-  .custom-slider::-webkit-slider-thumb:active,
-  .custom-slider::-moz-range-thumb:active {
-    background: #1054b1;
-  }
-
