+++ reactcode/src/router/index.js
@@ -1,24 +1,14 @@
 import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
-import ManualAnnotate from '../pages/manualAnnotate/index'
-import AutoAnnotate from '../pages/autoAnnotate/index'
-import Menu from '../App'
+import App from '../App'
 const router = createBrowserRouter([
     {
       path: '/',
       children: [
         {
             path: '/',
-            element: <Menu />
+            element: <App />
           },
-        {
-          path: '/annotate/auto',
-          element: <AutoAnnotate />
-        },
-        {
-          path: '/annotate/manual',
-          element: <ManualAnnotate />
-        },
       ]
     },
